major refactor, split up modules
